refactor: maintina proper spaces

+++ Assignment-11/ValueAndReferenceTypes/Program.cs
@@ -3,7 +3,7 @@
     internal class Program
     {
         /// <summary>
-        /// Function to modify the valuetype and referencetype variables.
+        /// Function to modify the value-type and reference-type variables.
         /// </summary>
         /// <param name="valueType">A value type parameter(int)</param>
         /// <param name="referenceType">A reference type </param>
@@ -29,8 +29,8 @@
         /// </summary>
         public static void PerformCalculations()
         {
-            int num1 = 11, num2 = 22, num3 = 33, num4 = 44, num5 = 555, num6 = 66, num7 = 77, num8 = 88, num9 = 99, num10=111;
-            int product = num1 * num2 * num3 * num4 * num5 * num6 * num7 * num8 * num9 * num10;
+            float num1 = 11, num2 = 22, num3 = 33, num4 = 44, num5 = 555, num6 = 66, num7 = 77, num8 = 88, num9 = 99, num10=111;
+            float product = num1 * num2 * num3 * num4 * num5 * num6 * num7 * num8 * num9 * num10;
         }
         public static void Main(string[] args)
         {
